[Update] GoogleAnalytics (3.0)
pod install started failing with the content of https://dl.google.com/dl/gaformobileapps/GoogleAnalyticsiOS.zip
After looking at it, it was because the content was updated to their new 3.0 SDK.
I wish they had the version in the URL so we would correctly target each version.
For now libGoogleAnalytics.a is not in the zip but the README says we can use libGoogleAnalyticsServices.a which is libGoogleAnalytics.a and the Tag manager lib combined. It's the only option for now.