Merge newly added TestFlightSDK-Beta back to TestFlightSDK
As Orta suggested there is no point of having a new pod TestFlightSDK-Beta that
in theory needs to be maintained indefinitely, instead the beta versions of the
TestFlight SDK can be distributed as a special versions of the existing
TestFlightSDK pod, with a version prefixed with 'b' e.g. "2.1.1b". This is
similar to what HockeySDK does.