Add Google AdMob SDK 6.8.0
The iOS SDK utilizes Apple's advertising identifier (IDFA). The SDK uses
IDFA under the guidelines laid out in the iOS developer program license
agreement. You must ensure you are in compliance with the iOS developer
program license agreement policies governing the use of this identifier.

This version of the SDK allows you to specify if you would like your app
to be treated as child-directed for purposes of the Children's Online
Privacy Protection Act (COPPA).