Convert all specifications to JSON.
This is the end of an era. We (the CocoaPods team) would like to thank
from the bottoms of our hearts, our very own Keith Smiley, The Legend.

With a deft flick of the thumb, @keithbsmiley has decided on the fate
of thousands of pull-requests. The community owes you its eternal
gratitude.

💖🎉💖

![Keith Smiley](https://avatars1.githubusercontent.com/u/283886)

These eyes have seen more pods than a south-american cacao farmer.