Unbreak BoringSSL.podspec for devices
🌈

Version 4.0 of the spec is broken when compiling for a device (i.e. arm
instead of simulator). Unfortunately, neither our CI tests nor `pod spec
lint` test on devices, so this slept through :(

For a number of reasons, upgrading the version for this fix isn’t
possible. Because the current spec is broken, and the problem is only in
the spec (not in the library sources), I hope this PR will be accepted.

This change was originally made in the Ruby version of the podspec (
https://github.com/grpc/grpc/pull/7264/commits/ccb184068d21758f905ecc3bfd42a2f9626dd5ee ).
That fix was confirmed to work both on simulator and devices. The spec
was then converted to JSON automatically, using the `pod ipc spec
BoringSSL.podspec > BoringSSL.podspec.json` command, to make sure no
errors were made in the translation.