Update Qualtrics.podspec to match M1 support.
According to the official release notes, version 2.4.0 added the actual support for arm64 based simulators. However, this podspec attempts to exclude the arm64 architecture after installing this CocoaPod which causes builds to fail on M1 Macs. We should be able to remove these flags from the podspec since the support is already there.